Move PromptChoice to beets.util module
And update imports that have been raising the deprecation warning.
